requirements.txt
# Python Version Required: 3.11+
# Tested with: 3.13.3
python>=3.11

# Core Framework
flet>=0.24.0

# Database & Authentication
supabase>=2.7.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# AI/LLM Integration
anthropic>=0.34.0  # Claude API
# openai>=1.0.0  # Alternative: OpenAI

# CSV & Data Handling
python-dateutil>=2.8.0

# Environment Management
python-dotenv>=1.0.0

# Optional: Plaid Bank Integration
# plaid-python>=24.0.0

# Charts & Visualization (Flet has built-in charts, but these are alternatives)
plotly>=5.18.0

# HTTP Requests
httpx>=0.27.0
requests>=2.31.0

# Utilities
pydantic>=2.0.0  # Data validation
